% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight_header}
\alias{highlight_header}
\alias{highlight_rows}
\alias{highlight_cols}
\alias{highlight_cells}
\alias{highlight_sum}
\alias{highlight_all}
\title{Highlight data frame output in terminal or markdown}
\usage{
highlight_header(x, styles)

highlight_rows(x, ..., styles)

highlight_cols(x, ..., styles)

highlight_cells(x, ..., styles)

highlight_sum(x, styles)

highlight_all(x, styles)
}
\arguments{
\item{x}{A data frame.}

\item{styles}{Style names or functions passed to `crayon::combine_styles()`.}

\item{...}{One or more unquoted expressions separated by commas, or integers
regarding the positions.}
}
\value{
A `tbl_hl`.
}
\description{
Highlight data frame output in terminal or markdown
}
